Redo cyanogenmod_tenderloin_defconfig.
Note that the one provided was unusable; it would crash between when adbd
starts and when the boot animation shows. The new config was ripped out of a
virgin, running CM11 nightly (2014-03-04).

* CONFIG_SWAP was turned on to make ZRAM useful.
* Since modules were disabled in the old config, and modules under Android are
generally a pain in the neck, zram was compiled into the kernel.
* CONFIG_CMDLINE and CONFIG_CMDLINE_EXTEND were used to permanently enable the
second zram device (the MSM8660 has two cores, so using a single ZRAM device would not givethe best performance.)

* All namespace options were enabled (multi-user features are supposed to benefit from them being available.)

* CONFIG_PANIC_TIMEOUT was lowered to 3, because many of the panics that still
persist happen while audio is being played, and cause a nasty sound that can't
be very good for the speakers
* CONFIG_LOG_BUF_SHIFT was maxed out at 2^21 (2 MiB) to make it easier to
 catch and hopefully fix any intermittent failures.
* Hung task and hard- and soft-lockup detection are now enabled.

* CONFIG_LOCALVERSION is now "-ninjaswag". Winning.

* CONFIG_JUMP_LABEL, CONFIG_SCHED_MC, was enabled to help performance.
* CONFIG_KSM was enabled because KitKat can use it. The ROM seems to know what
to do with it already - an option appears in Performance Settings to enable it.
* CONFIG_SND_USB_AUDIO was enabled to enable the connection of audio devices via USB-OTG, pending ROM support.

* ext[23] are now handled by the ext4 driver, which should improve performance when they are used.