tenderloin: add proper delay to allow panel/etc to power-on
Without this (added) delay, intermittent issues related to display
may occur. In particular:
  mdp4_lcdc_wait4dmap 402  TIMEOUT_
  mdp_enable_irq: MDP IRQ term-0x1 is already set, mask=101 irq=1
was observed intermittently during resume. Also, frequently coexisting
with the above timeout issue was persistent (until possibly cleared
by a future suspend/resume) display corruption where the leftmost
(if in landscape with home button to right) 6 or so column of pixels
will be swapped with the rightmost column of 6 or so pixels.